Make autoupdate --tags-only the default, add --bleeding-edge
